[Current workout page][Suggested exercises optimization]
Taking advantage of cached content by making a query every time there is only one letter in the string.

For example, a user is typing 'B'. Then all the exercises starting with B is fetched. Then the user types e ('Be'). Here since 'Be' results is a subset of ' B' results. It filters the data from the local cache and this will continue for subsequent chars until user deletes to whole string and starts from scratch.